requirements.txt
openai>=1.75.0
python-dotenv>=1.0.1
pydantic>=2.7.0
rich>=13.7.1
PyYAML>=6.0.1
pypdf>=4.0.0
flask>=3.0.0
werkzeug>=3.0.0



